+++ source/chip/STM32/STM32F4/include/distortos/chip/clocks.hpp
@@ -14,7 +14,7 @@
 
 #include "distortos/distortosConfiguration.h"
 
-#include <array>
+#include <cstdint>
 
 namespace distortos
 {
@@ -45,24 +45,45 @@ constexpr uint32_t minVcoOutFrequency {100000000};
 /// maximum allowed value for VCO output frequency, Hz
 constexpr uint32_t maxVcoOutFrequency {432000000};
 
-/// maximum allowed value for PLL output frequency, Hz
-/// [0] - in overdrive mode with voltage scale 1
-/// [1] - with voltage scale 1
-/// [2] - with voltage scale 2
-/// [3] - with voltage scale 3
+/// maximum allowed values for PLL output frequency, Hz
+/// [...][0] - over-drive disabled
+/// [...][1] - over-drive enabled
+/// [0][...] - voltage scale 1
+/// [1][...] - voltage scale 2
+/// [2][...] - voltage scale 3
 #if defined(CONFIG_CHIP_STM32F401)
-constexpr std::array<uint32_t, 4> maxPllOutFrequency {0, 0, 84000000, 60000000};
+constexpr uint32_t maxPllOutFrequencies[3][2]
+{
+		{84000000, 84000000},
+		{84000000, 84000000},
+		{60000000, 60000000},
+};
 #elif defined(CONFIG_CHIP_STM32F405) || defined(CONFIG_CHIP_STM32F407) || defined(CONFIG_CHIP_STM32F415) || \
 		defined(CONFIG_CHIP_STM32F417)
-constexpr std::array<uint32_t, 4> maxPllOutFrequency {0, 168000000, 144000000, 0};
+constexpr uint32_t maxPllOutFrequencies[3][2]
+{
+		{168000000, 168000000},
+		{144000000, 144000000},
+		{144000000, 144000000},
+};
 #elif defined(CONFIG_CHIP_STM32F410) || defined(CONFIG_CHIP_STM32F411) || defined(CONFIG_CHIP_STM32F412) || \
 		defined(CONFIG_CHIP_STM32F413) || defined(CONFIG_CHIP_STM32F423)
-constexpr std::array<uint32_t, 4> maxPllOutFrequency {0, 100000000, 84000000, 64000000};
+constexpr uint32_t maxPllOutFrequencies[3][2]
+{
+		{100000000, 100000000},
+		{84000000, 84000000},
+		{64000000, 64000000},
+};
 #else	// !defined(CONFIG_CHIP_STM32F401) && !defined(CONFIG_CHIP_STM32F405) && !defined(CONFIG_CHIP_STM32F407) &&
 		// !defined(CONFIG_CHIP_STM32F410) && !defined(CONFIG_CHIP_STM32F411) && !defined(CONFIG_CHIP_STM32F412) &&
 		// !defined(CONFIG_CHIP_STM32F413) && !defined(CONFIG_CHIP_STM32F415) && !defined(CONFIG_CHIP_STM32F417) &&
 		// !defined(CONFIG_CHIP_STM32F423) &&
-constexpr std::array<uint32_t, 4> maxPllOutFrequency {180000000, 168000000, 144000000, 120000000};
+constexpr uint32_t maxPllOutFrequencies[3][2]
+{
+		{168000000, 180000000},
+		{144000000, 144000000},
+		{120000000, 120000000},
+};
 #endif	// !defined(CONFIG_CHIP_STM32F401) && !defined(CONFIG_CHIP_STM32F405) && !defined(CONFIG_CHIP_STM32F407) &&
 		// !defined(CONFIG_CHIP_STM32F410) && !defined(CONFIG_CHIP_STM32F411) && !defined(CONFIG_CHIP_STM32F412) &&
 		// !defined(CONFIG_CHIP_STM32F413) && !defined(CONFIG_CHIP_STM32F415) && !defined(CONFIG_CHIP_STM32F417) &&
@@ -71,35 +92,39 @@ constexpr std::array<uint32_t, 4> maxPllOutFrequency {180000000, 168000000, 1440
 /// maximum allowed value for PLL "Q" output frequency, Hz
 constexpr uint32_t maxPllqOutFrequency {48000000};
 
-/// maximum allowed APB1 (low speed) frequency, Hz
+/// maximum allowed APB1 (low speed) frequencies, Hz
+/// [0] - over-drive disabled
+/// [1] - over-drive enabled
 #if defined(CONFIG_CHIP_STM32F401) || defined(CONFIG_CHIP_STM32F405) || defined(CONFIG_CHIP_STM32F407) || \
 		defined(CONFIG_CHIP_STM32F415) || defined(CONFIG_CHIP_STM32F417)
-constexpr uint32_t maxApb1Frequency {42000000};
+constexpr uint32_t maxApb1Frequencies[2] {42000000, 42000000};
 #elif defined(CONFIG_CHIP_STM32F410) || defined(CONFIG_CHIP_STM32F411) || defined(CONFIG_CHIP_STM32F412) || \
 		defined(CONFIG_CHIP_STM32F413) || defined(CONFIG_CHIP_STM32F423)
-constexpr uint32_t maxApb1Frequency {50000000};
+constexpr uint32_t maxApb1Frequencies[2] {50000000, 50000000};
 #else	// !defined(CONFIG_CHIP_STM32F401) && !defined(CONFIG_CHIP_STM32F405) && !defined(CONFIG_CHIP_STM32F407) &&
 		// !defined(CONFIG_CHIP_STM32F410) && !defined(CONFIG_CHIP_STM32F411) && !defined(CONFIG_CHIP_STM32F412) &&
 		// !defined(CONFIG_CHIP_STM32F413) && !defined(CONFIG_CHIP_STM32F415) && !defined(CONFIG_CHIP_STM32F417) &&
 		// !defined(CONFIG_CHIP_STM32F423)
-constexpr uint32_t maxApb1Frequency {45000000};
+constexpr uint32_t maxApb1Frequencies[2] {42000000, 45000000};
 #endif	// !defined(CONFIG_CHIP_STM32F401) && !defined(CONFIG_CHIP_STM32F405) && !defined(CONFIG_CHIP_STM32F407) &&
 		// !defined(CONFIG_CHIP_STM32F410) && !defined(CONFIG_CHIP_STM32F411) && !defined(CONFIG_CHIP_STM32F412) &&
 		// !defined(CONFIG_CHIP_STM32F413) && !defined(CONFIG_CHIP_STM32F415) && !defined(CONFIG_CHIP_STM32F417) &&
 		// !defined(CONFIG_CHIP_STM32F423)
 
-/// maximum allowed APB2 (high speed) frequency, Hz
+/// maximum allowed APB2 (high speed) frequencies, Hz
+/// [0] - over-drive disabled
+/// [1] - over-drive enabled
 #if defined(CONFIG_CHIP_STM32F401) || defined(CONFIG_CHIP_STM32F405) || defined(CONFIG_CHIP_STM32F407) || \
 		defined(CONFIG_CHIP_STM32F415) || defined(CONFIG_CHIP_STM32F417)
-constexpr uint32_t maxApb2Frequency {84000000};
+constexpr uint32_t maxApb2Frequencies[2] {84000000, 84000000};
 #elif defined(CONFIG_CHIP_STM32F410) || defined(CONFIG_CHIP_STM32F411) || defined(CONFIG_CHIP_STM32F412) || \
 	defined(CONFIG_CHIP_STM32F413) || defined(CONFIG_CHIP_STM32F423)
-constexpr uint32_t maxApb2Frequency {100000000};
+constexpr uint32_t maxApb2Frequencies[2] {100000000, 100000000};
 #else	// !defined(CONFIG_CHIP_STM32F401) && !defined(CONFIG_CHIP_STM32F405) && !defined(CONFIG_CHIP_STM32F407) &&
 		// !defined(CONFIG_CHIP_STM32F410) && !defined(CONFIG_CHIP_STM32F411) && !defined(CONFIG_CHIP_STM32F412) &&
 		// !defined(CONFIG_CHIP_STM32F413) && !defined(CONFIG_CHIP_STM32F415) && !defined(CONFIG_CHIP_STM32F417) &&
 		// !defined(CONFIG_CHIP_STM32F423)
-constexpr uint32_t maxApb2Frequency {90000000};
+constexpr uint32_t maxApb2Frequencies[2] {84000000, 90000000};
 #endif	// !defined(CONFIG_CHIP_STM32F401) && !defined(CONFIG_CHIP_STM32F405) && !defined(CONFIG_CHIP_STM32F407) &&
 		// !defined(CONFIG_CHIP_STM32F410) && !defined(CONFIG_CHIP_STM32F411) && !defined(CONFIG_CHIP_STM32F412) &&
 		// !defined(CONFIG_CHIP_STM32F413) && !defined(CONFIG_CHIP_STM32F415) && !defined(CONFIG_CHIP_STM32F417) &&
@@ -115,19 +140,31 @@ static_assert(CONFIG_CHIP_STM32F4_PWR_VOLTAGE_SCALE_MODE == 1, "Over-drive mode 
 static_assert(CONFIG_CHIP_STM32F4_VDD_MV >= 2100,
 		"Over-drive mode must not be enabled when supply voltage is below 2.1V!");
 
-/// voltage scale index for \a maxPllOutFrequency array
-constexpr uint8_t voltageScaleIndex {0};
+/// over-drive index for \a maxPllOutFrequencies array (maxPllOutFrequencies[voltageScaleIndex][overDriveIndex])
+constexpr uint8_t overDriveIndex {1};
 
 #else	// !(defined(CONFIG_CHIP_STM32F427) || defined(CONFIG_CHIP_STM32F429) || defined(CONFIG_CHIP_STM32F43) ||
 		// defined(CONFIG_CHIP_STM32F446) || defined(CONFIG_CHIP_STM32F469) || defined(CONFIG_CHIP_STM32F479)) ||
 		// !defined(CONFIG_CHIP_STM32F4_PWR_OVER_DRIVE_ENABLE)
 
-/// voltage scale index for \a maxPllOutFrequency array
-constexpr uint8_t voltageScaleIndex {CONFIG_CHIP_STM32F4_PWR_VOLTAGE_SCALE_MODE};
+/// over-drive index for \a maxPllOutFrequencies array (maxPllOutFrequencies[voltageScaleIndex][overDriveIndex])
+constexpr uint8_t overDriveIndex {0};
 
 #endif	// !(defined(CONFIG_CHIP_STM32F427) || defined(CONFIG_CHIP_STM32F429) || defined(CONFIG_CHIP_STM32F43) ||
 		// defined(CONFIG_CHIP_STM32F446) || defined(CONFIG_CHIP_STM32F469) || defined(CONFIG_CHIP_STM32F479)) ||
 		// !defined(CONFIG_CHIP_STM32F4_PWR_OVER_DRIVE_ENABLE)
+
+/// voltage scale index for \a maxPllOutFrequencies array (maxPllOutFrequencies[voltageScaleIndex][overDriveIndex])
+constexpr uint8_t voltageScaleIndex {CONFIG_CHIP_STM32F4_PWR_VOLTAGE_SCALE_MODE - 1};
+
+/// maximum allowed value for PLL output frequency, Hz
+constexpr uint32_t maxPllOutFrequency {maxPllOutFrequencies[voltageScaleIndex][overDriveIndex]};
+
+/// maximum allowed APB1 (low speed) frequency, Hz
+constexpr uint32_t maxApb1Frequency {maxApb1Frequencies[overDriveIndex]};
+
+/// maximum allowed APB2 (high speed) frequency, Hz
+constexpr uint32_t maxApb2Frequency {maxApb2Frequencies[overDriveIndex]};
 
 #ifdef CONFIG_CHIP_STM32F4_RCC_PLL_ENABLE
 
@@ -153,7 +190,7 @@ static_assert(minVcoOutFrequency <= vcoOutFrequency && vcoOutFrequency <= maxVco
 /// PLL output frequency, Hz
 constexpr uint32_t pllOutFrequency {vcoOutFrequency / CONFIG_CHIP_STM32F4_RCC_PLLP};
 
-static_assert(pllOutFrequency <= maxPllOutFrequency[voltageScaleIndex], "Invalid PLL output frequency!");
+static_assert(pllOutFrequency <= maxPllOutFrequency, "Invalid PLL output frequency!");
 
 /// PLL "Q" output frequency, Hz
 constexpr uint32_t pllqOutFrequency {vcoOutFrequency / CONFIG_CHIP_STM32F4_RCC_PLLQ};
@@ -186,6 +223,12 @@ constexpr uint32_t sysclkFrequency {pllrOutFrequency};
 
 /// SYSCLK frequency, Hz
 constexpr uint32_t sysclkFrequency {CONFIG_CHIP_STM32F4_RCC_SYSCLK_FREQUENCY};
+
+/// maximum allowed APB1 (low speed) frequency, Hz
+constexpr uint32_t maxApb1Frequency {maxApb1Frequencies[1]};
+
+/// maximum allowed APB2 (high speed) frequency, Hz
+constexpr uint32_t maxApb2Frequency {maxApb2Frequencies[1]};
 
 #endif	// !def CONFIG_CHIP_STM32F4_STANDARD_CLOCK_CONFIGURATION_ENABLE